Committed latest from long time ago

+++ admin_pages/views.py
@@ -1,19 +1,16 @@
-from django.shortcuts import redirect, render
-from django.http import HttpResponse, HttpResponseRedirect, Http404
+from django.shortcuts import render
+from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.contrib.auth.models import User
 from django.contrib.auth.decorators import login_required
 from django.core.files.storage import FileSystemStorage
-from django.core.exceptions import MultipleObjectsReturned
-from django.views.decorators.cache import never_cache
-from django.forms import ValidationError
 
 from dashboard import settings
-from dashboard import context_processors
+from site_pages.models import Home
+from site_pages.forms import HomeForm
 from .models import *
 from .forms import *
 
-import json
 import os
 
 
@@ -88,6 +85,20 @@ def manage_site_look(request):
 
 
 @login_required
+def manage_home(request):
+    home = Home.objects.get(id=1)
+    if request.method != 'POST':
+        form = HomeForm(instance=home)
+    else:
+        form = HomeForm(instance=home, data=request.POST)
+        if form.is_valid():
+            form.save()
+            return HttpResponseRedirect(reverse('admin_pages:admin_index'))
+    context = { 'form': form }
+    return render(request, 'admin_pages/manage_home/manage_home.html', context)
+
+
+@login_required
 def manage_user_profile(request):
     profile = UserProfile.objects.get(user=User.objects.get(id=request.user.id))
     if request.method != 'POST':
@@ -100,7 +111,8 @@ def manage_user_profile(request):
 
             if len(request.FILES) != 0:
                 fs_storage = FileSystemStorage(location=STATIC_IMGS_DIR)
-                if profile.profile_img_name != '':
+                if ((profile.profile_img_name != '') 
+                    and (profile.profile_img_name != 'profile-default.gif')): # Prevent deletion of default image.
                     try:
                         os.remove(os.path.join(STATIC_IMGS_DIR,
                             profile.profile_img_name))
@@ -121,7 +133,7 @@ def manage_user_profile(request):
 def manage_users(request):
     """Provides the user a form template to facilitate creation and deletion of a
     User admin (ish) account. The user is considered an admin for access to the 
-    Admin Pages and API but is not a true Django admin which is unneccessary since
+    Admin Pages and API but is not a true Django admin which is unnecessary since
     there are no admin-registered models.
 
     Parameters:
@@ -137,3 +149,81 @@ def manage_users(request):
         return render(request, 'admin_pages/manage_users/manage_users.html', context)
     else:
         return HttpResponseRedirect(reverse('admin_pages:admin_index'))
+
+@login_required
+def manage_pet_sales_opts(request):
+    pet_sales_opts = PetSalesOpts.objects.get(id=1)
+    if request.method != 'POST':
+        form = PetSalesOptsForm(instance=pet_sales_opts)
+    else:
+        form = PetSalesOptsForm(instance=pet_sales_opts, data=request.POST)
+        if form.is_valid():
+            form.save()
+            return HttpResponseRedirect(reverse('admin_pages:admin_index'))
+    context = { 'form': form }
+    return render(request, 'admin_pages/manage_metrics_display_opts/manage_petsales.html', context)
+
+@login_required
+def manage_product_sales_opts(request):
+    product_sales_opts = ProductSalesOpts.objects.get(id=1)
+    if request.method != 'POST':
+        form = ProductSalesOptsForm(instance=product_sales_opts)
+    else:
+        form = ProductSalesOptsForm(instance=product_sales_opts, data=request.POST)
+        if form.is_valid():
+            form.save()
+            return HttpResponseRedirect(reverse('admin_pages:admin_index'))
+    context = { 'form': form } 
+    return render(request, 'admin_pages/manage_metrics_display_opts/manage_productsales.html', context)
+
+@login_required
+def manage_site_visits_opts(request):
+    site_visits_opts = SiteVisitsOpts.objects.get(id=1)
+    if request.method != 'POST':
+        form = SiteVisitsOptsForm(instance=site_visits_opts)
+    else:
+        form = SiteVisitsOptsForm(instance=site_visits_opts, data=request.POST)
+        if form.is_valid():
+            form.save()
+            return HttpResponseRedirect(reverse('admin_pages:admin_index'))
+    context = { 'form': form } 
+    return render(request, 'admin_pages/manage_metrics_display_opts/manage_sitevisits.html', context)
+
+@login_required
+def manage_pages_viewed_opts(request):
+    pages_viewed_opts = PagesViewedOpts.objects.get(id=1)
+    if request.method != 'POST':
+        form = PagesViewedOptsForm(instance=pages_viewed_opts)
+    else:
+        form = PagesViewedOptsForm(instance=pages_viewed_opts, data=request.POST)
+        if form.is_valid():
+            form.save()
+            return HttpResponseRedirect(reverse('admin_pages:admin_index'))
+    context = { 'form': form } 
+    return render(request, 'admin_pages/manage_metrics_display_opts/manage_pagesviewed.html', context)
+
+@login_required
+def manage_users_opts(request):
+    users_opts = UsersOpts.objects.get(id=1)
+    if request.method != 'POST':
+        form = UsersOptsForm(instance=users_opts)
+    else:
+        form = UsersOptsForm(instance=users_opts, data=request.POST)
+        if form.is_valid():
+            form.save()
+            return HttpResponseRedirect(reverse('admin_pages:admin_index'))
+    context = { 'form': form } 
+    return render(request, 'admin_pages/manage_metrics_display_opts/manage_users.html', context)
+
+@login_required
+def manage_day_and_time_opts(request):
+    day_and_time_opts = DayAndTimeOpts.objects.get(id=1)
+    if request.method != 'POST':
+        form = DayAndTimeOptsForm(instance=day_and_time_opts)
+    else:
+        form = DayAndTimeOptsForm(instance=day_and_time_opts, data=request.POST)
+        if form.is_valid():
+            form.save()
+            return HttpResponseRedirect(reverse('admin_pages:admin_index'))
+    context = { 'form': form }
+    return render(request, 'admin_pages/manage_metrics_display_opts/manage_dayandtime.html', context)
